+++ sdk/eventgrid/azure-eventgrid/samples/consume_samples/consume_cloud_custom_data_sample.py
@@ -28,7 +28,14 @@ cloud_custom_bytes = str(cloud_custom_string).encode("utf-8")
 
 client = EventGridDeserializer()
 deserialized_dict_event = client.deserialize_cloud_events(cloud_custom_dict)
-print(deserialized_dict_event)
+
+# to return raw json, we use the data param
+print(deserialized_dict_event.data)
+print(type(deserialized_dict_event.data)) # this is a dictionary
+
+# use system_event_data to get a strongly typed model
+print(deserialized_dict_event.system_event_data)
+print(type(deserialized_dict_event.system_event_data)) # this is a strongly typed model
 
 deserialized_str_event = client.deserialize_cloud_events(cloud_custom_string)
 print(deserialized_str_event)