+++ sdk/eventgrid/azure-eventgrid/tests/test_eg_consumer.py
@@ -37,40 +37,49 @@ class EventGridDeserializerTests(AzureMgmtTestCase):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_cloud_events(cloud_storage_dict)
         assert deserialized_event.__class__ == CloudEvent
-        assert deserialized_event.data.__class__ == StorageBlobCreatedEventData
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data.__class__ == StorageBlobCreatedEventData
 
     def test_eg_consumer_cloud_storage_string(self, **kwargs):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_cloud_events(cloud_storage_string)
         assert deserialized_event.__class__ == CloudEvent
-        assert deserialized_event.data.__class__ == StorageBlobCreatedEventData
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data.__class__ == StorageBlobCreatedEventData
 
     def test_eg_consumer_cloud_storage_bytes(self, **kwargs):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_cloud_events(cloud_storage_bytes)
         assert deserialized_event.__class__ == CloudEvent
-        assert deserialized_event.data.__class__ == StorageBlobCreatedEventData
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data.__class__ == StorageBlobCreatedEventData
     
 
     def test_eg_consumer_cloud_custom_dict(self, **kwargs):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_cloud_events(cloud_custom_dict)
         assert deserialized_event.__class__ == CloudEvent
-        assert deserialized_event.data is None
+        assert deserialized_event.data == {"team": "event grid squad"}
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data is None
 
 
     def test_eg_consumer_cloud_custom_string(self, **kwargs):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_cloud_events(cloud_custom_string)
         assert deserialized_event.__class__ == CloudEvent
-        assert deserialized_event.data is None
+        assert deserialized_event.data == {"team": "event grid squad"}
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data is None
 
 
     def test_eg_consumer_cloud_custom_bytes(self, **kwargs):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_cloud_events(cloud_custom_bytes)
         assert deserialized_event.__class__ == CloudEvent
-        assert deserialized_event.data is None
+        assert deserialized_event.data == {"team": "event grid squad"}
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data is None
     
     # EG Event tests
 
@@ -78,39 +87,48 @@ class EventGridDeserializerTests(AzureMgmtTestCase):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_eventgrid_events(eg_storage_dict)
         assert deserialized_event.__class__ == EventGridEvent
-        assert deserialized_event.data.__class__ == StorageBlobCreatedEventData
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data.__class__ == StorageBlobCreatedEventData
 
 
     def test_eg_consumer_eg_storage_string(self, **kwargs):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_eventgrid_events(eg_storage_string)
         assert deserialized_event.__class__ == EventGridEvent
-        assert deserialized_event.data.__class__ == StorageBlobCreatedEventData
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data.__class__ == StorageBlobCreatedEventData
 
 
     def test_eg_consumer_eg_storage_bytes(self, **kwargs):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_eventgrid_events(eg_storage_bytes)
         assert deserialized_event.__class__ == EventGridEvent
-        assert deserialized_event.data.__class__ == StorageBlobCreatedEventData
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data.__class__ == StorageBlobCreatedEventData
     
 
     def test_eg_consumer_eg_custom_dict(self, **kwargs):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_eventgrid_events(eg_custom_dict)
         assert deserialized_event.__class__ == EventGridEvent
-        assert deserialized_event.data is None
+        assert deserialized_event.data == {"team": "event grid squad"}
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data is None
 
 
     def test_eg_consumer_eg_custom_string(self, **kwargs):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_eventgrid_events(eg_custom_string)
         assert deserialized_event.__class__ == EventGridEvent
-        assert deserialized_event.data is None
+        assert deserialized_event.data == {"team": "event grid squad"}
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data is None
 
 
     def test_eg_consumer_eg_custom_bytes(self, **kwargs):
         client = EventGridDeserializer()
         deserialized_event = client.deserialize_eventgrid_events(eg_custom_bytes)
         assert deserialized_event.__class__ == EventGridEvent
-        assert deserialized_event.data is None
+        assert deserialized_event.data == {"team": "event grid squad"}
+        assert deserialized_event.data.__class__ == dict
+        assert deserialized_event.system_event_data is None